Move net/node to root folder of nkn
1. move net/node to root folder of nkn
2. delete net folder
3. modify all imports to net/node
