+++ api/common/common.go
@@ -2,7 +2,7 @@ package common
 
 import (
 	"github.com/nknorg/nkn/errors"
-	"github.com/nknorg/nkn/net/node"
+	"github.com/nknorg/nkn/node"
 	"github.com/nknorg/nkn/vault"
 )
