+++ node/hashcache.go
