+++ node/syncblock.go
@@ -9,7 +9,7 @@ import (
 	"github.com/gogo/protobuf/proto"
 	"github.com/nknorg/nkn/blockchain"
 	"github.com/nknorg/nkn/common"
-	"github.com/nknorg/nkn/net/node/consequential"
+	"github.com/nknorg/nkn/node/consequential"
 	"github.com/nknorg/nkn/pb"
 	"github.com/nknorg/nkn/types"
 	"github.com/nknorg/nkn/util/config"
